+++ packages/node/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add support to chainTypes, allow user use hasher name string in chainTypes
+
 ## [4.4.2] - 2024-05-20
 ### Fixed
 - Fix load chain types missing some builtins modules